Move stuff to env variables
